codeflex-parcial1-sw_api/requirements.txt
Django==4.2.16

djangorestframework==3.15.2

uvicorn[standard]

channels==4.1.0
channels-redis==4.2.0
django-redis==5.4.0

whitenoise==6.8.2

django-environ==0.9.0

pillow==11.0.0

django-ckeditor==6.7.1

psycopg2-binary

django-cors-headers==4.6.0

djoser==2.3.1

djangorestframework-simplejwt==5.3.1

cryptography==41.0.7

bleach==6.2.0

argon2-cffi==21.3.0

argon2-cffi-bindings==21.2.0